change a type of a variable from int to size_t

+++ modules/stitching/src/matchers.cpp
@@ -713,7 +713,7 @@ void BestOf2NearestOfNeighborMatcher::operator ()(const std::vector<ImageFeature
 
     std::vector<std::pair<int,int> > near_pairs;
     for (int i = 0; i < num_images - 1; ++i)
-        for (int j = 0; j < adjacencies_list_.size(); j++)
+        for (size_t j = 0; j < adjacencies_list_.size(); j++)
             if (features[i].keypoints.size() > 0 && features[j].keypoints.size() > 0 && mask_(i, j))
                 near_pairs.push_back(std::make_pair(i, j));
